+++ src/components/file-upload/z-file/index.tsx
@@ -9,11 +9,10 @@ import {
   Listen,
   Prop,
 } from "@stencil/core";
-import { DeviceEnum, TooltipPosition, ZChipType } from "../../../beans";
+import { TooltipPosition, ZChipType } from "../../../beans";
 import { tabletBreakpoint } from "../../../constants/breakpoints";
-import { getDevice } from "../../../utils/utils";
 
-import './index';
+import "./index";
 
 @Component({
   tag: "z-file",
@@ -25,12 +24,12 @@ export class ZFile {
   private icon: HTMLElement;
 
   private ellipsis?: HTMLSpanElement;
-  
+
   @Prop() fileNumber: number;
 
   @State() allowTooltip: boolean = false;
 
-  @State() tooltipVisible: boolean = false
+  @State() tooltipVisible: boolean = false;
 
   @Element() el: HTMLElement;
 
@@ -41,14 +40,19 @@ export class ZFile {
     this.el.remove();
   }
 
-  @Listen('mouseover')
+  @Listen("mouseover")
   onMouseOver() {
     this.tooltipVisible = true;
   }
 
-  @Listen('mouseleave')
+  @Listen("mouseleave")
   onMouseLeave() {
     this.tooltipVisible = false;
+  }
+
+  @Listen("interactiveIconClick")
+  onInteractiveIconClick() {
+    this.removeFileHandler();
   }
 
   componentDidLoad() {
@@ -59,6 +63,7 @@ export class ZFile {
   }
 
   elementHasEllipsis(): boolean {
+    console.log(this.ellipsis, this.ellipsis.offsetWidth, this.ellipsis.scrollWidth);
     return this.ellipsis.offsetWidth < this.ellipsis.scrollWidth;
   }
 
@@ -74,34 +79,17 @@ export class ZFile {
             <span class="body-5 tootip-content">{this.ellipsis.innerText}</span>
           </z-tooltip>
         )}
-        <z-chip id={`chip-${this.fileNumber}`} filter type={ZChipType.default}>
-          <div class="chip-content">
+        <z-chip
+          id={`chip-${this.fileNumber}`}
+          interactiveIcon="multiply-circled"
+          type={ZChipType.default}
+        >
             <span
               ref={(el) => (this.ellipsis = el as HTMLSpanElement)}
               tabIndex={-1}
-              class={{
-                "body-3-sb": getDevice() == DeviceEnum.desktop,
-                "body-2-sb": getDevice() !== DeviceEnum.desktop,
-              }}
             >
               <slot />
             </span>
-            <z-icon
-              aria-label="Elimina file"
-              tabIndex={0}
-              onClick={() => this.removeFileHandler()}
-              onKeyPress={(e) => {
-                if (e.code == "Space" || e.code == "Enter") {
-                  e.preventDefault();
-                  this.removeFileHandler();
-                }
-              }}
-              name="multiply-circled"
-              height={getDevice() !== DeviceEnum.desktop ? 20 : 15}
-              width={getDevice() !== DeviceEnum.desktop ? 20 : 15}
-              ref={(val) => (this.icon = val as HTMLElement)}
-            />
-          </div>
         </z-chip>
       </Host>
     );